requirements.txt
telethon
binance-futures-connector
flask
requests
python-dotenv